% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{adj_mat}
\alias{adj_mat}
\title{Covert data matrix to adjacency matrix}
\usage{
adj_mat(x, method = "euclidian")
}
\arguments{
\item{x}{a list obtained by the \code{mnda_embedding_2layer()} function.}

\item{method}{method for adjusting p-value (including methods on \code{p.adjust.methods}).
If set to "none" (default), no adjustment will be performed.}
}
\description{
Covert data matrix to adjacency matrix
}
\examples{
X = gaussian_clusters()$X
Adj = adj_mat(X, method = "euclidian")

}
