% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CC_with_robScore_functions.R
\name{consensus_matrix}
\alias{consensus_matrix}
\title{Calculate consensus matrix for data perturbation consensus clustering}
\usage{
consensus_matrix(
  X,
  max.cluster = 5,
  resample.ratio = 0.7,
  max.itter = 100,
  clustering.method = "hclust",
  no.cores = 1,
  adj.conv = TRUE
)
}
\arguments{
\item{X}{adjacency matrix a Nsample x Nsample}

\item{max.cluster}{maximum number of clusters}

\item{resample.ratio}{the data ratio to use at each itteration.}

\item{max.itter}{maximum number of itterations at each \code{max.cluster}}

\item{clustering.method}{base clustering method: \code{c("hclust", "spectral", "pam")}}
}
\description{
Calculate consensus matrix for data perturbation consensus clustering
}
\details{
performs data perturbation consensus clustering and obtain consensus matrix
Monti et al. (2003) consensus clustering algorithm
}
\examples{
consensus_matrix(X)

}
