% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusCluster-package.R
\docType{package}
\name{ConsensusCluster-package}
\alias{ConsensusCluster}
\alias{ConsensusCluster-package}
\title{ConsensusCluster: Iterative Consensus Clustering}
\description{
This package implements iterative consensus clustering with multiple
distance metrics using efficient C++ code.
}
