% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_generation_functions.R
\name{divide_interval_int}
\alias{divide_interval_int}
\title{??}
\usage{
divide_interval_int(x, n)
}
\arguments{
\item{x}{x}

\item{n}{n}
}
\description{
??
}
\examples{
divide_interval_int(10,3)

}
