% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{indicator_matrix}
\alias{indicator_matrix}
\title{Build indicator matrix}
\usage{
indicator_matrix(clusters)
}
\arguments{
\item{clusters}{a vector of clusterings. Zero elements mean that the sample was absent
during clustering}
}
\description{
Build indicator matrix
}
\details{
Indicator matrix (I) is a binary matrix N-by-N
 I[i,j] = 1 if sample i and j co-exist for clustering
 ref: Monti et al. (2003) "Consensus Clustering: A Resampling-Based Method for
 Class Discovery and Visualization of Gene Expression Microarray Data", Machine Learning
}
