% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CC_without_robScore_functions.R
\name{multi_pam_gen}
\alias{multi_pam_gen}
\title{Multiple PAM (K-medoids) generation}
\usage{
multi_pam_gen(
  X,
  rep = 10,
  range.k = c(2, 5),
  is.distance = FALSE,
  method = "random"
)
}
\arguments{
\item{X}{input data Nsample x Nfeatures or distance matrix.}

\item{rep}{number of repeats}

\item{range.k}{vector of minimum and maximum values for k \code{c(min, max)}}

\item{is.distance}{binary balue indicating if the input \code{X} is distance}

\item{method}{method for the choice of k at each repeat \code{c("random", "silhouette")}}
}
\description{
Multiple PAM (K-medoids) generation
}
\details{
At each repeat, k is selected randomly or based on the best silhouette width from a discrete uniform distribution between range.k[1] and range.k[2].
Then PAM clustering is applied and result is returned.
}
\examples{
multi_pam_gen(X)


}
