% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CC_with_robScore_functions.R
\name{multiview_consensus_matrix}
\alias{multiview_consensus_matrix}
\title{Calculate consensus matrix for multi-data consensus clustering}
\usage{
multiview_consensus_matrix(
  X,
  max.cluster = 5,
  sample.set = NA,
  clustering.method = "hclust",
  no.cores = 1,
  adj.conv = TRUE
)
}
\arguments{
\item{X}{list of adjacency matrices for different cohorts (or views).}

\item{max.cluster}{maximum number of clusters}

\item{sample.set}{vector of samples the clustering is being applied on. \code{sample.set} can be names or indices.
if \code{sample.set} is \code{NA}, it considers that all the datasets have the same samples with the same order.}

\item{clustering.method}{base clustering method: \code{c("hclust", "spectral", "pam")}}

\item{no.cores}{number of cores}

\item{adj.conv}{binary value to apply soft threshold (default=\code{TRUE})}
}
\description{
Calculate consensus matrix for multi-data consensus clustering
}
\details{
performs multi-data consensus clustering and obtain consensus matrix
Monti et al. (2003) consensus clustering algorithm
}
\examples{
# multiview_consensus_matrix()

}
