% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CC_without_robScore_functions.R
\name{multiview_kmeans_gen}
\alias{multiview_kmeans_gen}
\title{Multiview K-means generation}
\usage{
multiview_kmeans_gen(X, rep = 10, range.k = c(2, 5), method = "random")
}
\arguments{
\item{X}{List of input data matrices of Sample x feature. The length of \code{X} is equal to Nviews}

\item{rep}{number of repeats}

\item{range.k}{vector of minimum and maximum values for k \code{c(min, max)}}

\item{method}{method for the choice of k at each repeat \code{c("random", "silhouette")}}
}
\description{
Multiview K-means generation
}
\details{
At each repeat, k is selected randomly or based on the best silhouette width from a discrete uniform distribution between range.k[1] and range.k[2].
Then k-means clustering is applied and result is returned.
}
\examples{
multiview_kmeans_gen(X)


}
