% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CC_without_robScore_functions.R
\name{multiview_pam_gen}
\alias{multiview_pam_gen}
\title{Multiview PAM (K-medoids) generation}
\usage{
multiview_pam_gen(
  X,
  func,
  rep = 10,
  param,
  method = "random",
  is.distance = FALSE,
  sample.set = NA
)

multiview_pam_gen(
  X,
  func,
  rep = 10,
  param,
  method = "random",
  is.distance = FALSE,
  sample.set = NA
)
}
\arguments{
\item{X}{List of input data matrices of Sample x feature or distance matrices.
The length of \code{X} is equal to Nviews}

\item{func}{custom function that accepts \code{X} and a parameter that return a vector of clusterings.
\code{cluster_func <- function(X, param)}}

\item{rep}{number of repeats}

\item{param}{vector of parameters}

\item{method}{method for the choice of k at each repeat \code{c("random", "silhouette")}}

\item{is.distance}{binary balue indicating if the input \code{X[i]} is distance}

\item{sample.set}{vector of samples the clustering is being applied on. can be names or indices.
if \code{sample.set} is \code{NA}, it considers all the datasets have the same samples with the same order}

\item{range.k}{vector of minimum and maximum values for k \code{c(min, max)}}
}
\description{
Multiview PAM (K-medoids) generation

Multiview cluster generation
}
\details{
At each repeat, k is selected randomly or based on the best silhouette width from a discrete uniform distribution between range.k[1] and range.k[2].
Then PAM clustering is applied and result is returned.

At each repeat, k is selected randomly or based on the best silhouette width from a discrete uniform distribution between range.k[1] and range.k[2].
Then clustering is applied and result is returned.
}
\examples{
multiview_kmeans_gen(X)


multiview_kmeans_gen(X)


}
