% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_cluster.R
\name{print.consensus_cluster}
\alias{print.consensus_cluster}
\title{Print method for consensus_cluster objects}
\usage{
\method{print}{consensus_cluster}(x, ...)
}
\arguments{
\item{x}{A consensus_cluster object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for consensus_cluster objects
}
