% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{coCluster_matrix}
\alias{coCluster_matrix}
\title{Calculate the  Co-cluster matrix for a set of given set of clustering results.}
\usage{
coCluster_matrix(X)
}
\arguments{
\item{X}{clustering matrix of Nsamples x Nclusterings.
Zero elements are are considered as unclustered samples}
}
\description{
Calculate the  Co-cluster matrix for a set of given set of clustering results.
}
\examples{
Clustering = cbind(c(1,1,1,2,2,2),
                   c(1,1,2,1,2,2))
coCluster_matrix(Clustering)

}
