% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{connectivity_matrix}
\alias{connectivity_matrix}
\title{Build connectivity matrix}
\usage{
connectivity_matrix(clusters)
}
\arguments{
\item{clusters}{a vector of clusterings. Zero elements mean that the sample was absent
during clustering}
}
\value{
Connectivity matrix
}
\description{
Build connectivity matrix
}
\details{
Connectivity matrix (M) is a binary matrix N-by-N
 M[i,j] = 1 if sample i and j are in the same cluster
 ref: Monti et al. (2003) "Consensus Clustering: A Resampling-Based Method for
 Class Discovery and Visualization of Gene Expression Microarray Data", Machine Learning
}
\examples{
con_mat = connectivity_matrix(c(1,1,1,2,2,2))

}
